+++ ngs_mapper/tests/test_trim_reads.py
@@ -27,6 +27,35 @@ class TestTrimReadsInDir(TrimBase):
     def setUp( self ):
         super( TestTrimReadsInDir, self ).setUp()
 
+    @attr('current')
+    @patch('ngs_mapper.trim_reads.os', MagicMock())
+    @patch('__builtin__.open')
+    @patch('ngs_mapper.trim_reads.data')
+    def test_skips_miseq_index_reads(self, mdata, mopen):
+        reads = mdata.reads_by_plat.return_value = {
+            'Sanger': [
+                '1710_F2824_2014_01_14_Den4_Den4_1274_A01.ab1',
+                '1710_F2824_2014_01_14_Den4_Den4_1274_A01.fastq'
+            ],
+            'MiSeq': [
+                ('foo_S01_L001_R1_001_2001_01_01.fastq','foo_S01_L001_R2_001_2001_01_01.fastq'),
+                'foo_S01_L001_I1_001_2001_01_01.fastq','foo_S01_L001_I2_001_2001_01_01.fastq',
+            ],
+            'foo': [
+                'bar.fastq',
+                'baz.sff'
+            ]
+        }
+    
+        with patch('ngs_mapper.trim_reads.trim_read') as mtrim_read:
+            self._C('/path/to/reads', 20, '/path/to/outdir')
+            #_call = call(reads['Sanger'][1], 20, '/path/to/outdir/{0}'.format(basename(reads['Sanger'][1])), head_crop=0)
+            for call in mtrim_read.call_args_list:
+                assert_not_in(
+                    '_I', call[0][0],
+                    'Did not skip MiSeq Index {0}'.format(call[0][0])
+                )
+
     @patch('ngs_mapper.trim_reads.os', MagicMock())
     @patch('__builtin__.open')
     @patch('ngs_mapper.trim_reads.data')