+++ ngs_mapper/tests/unit/__init__.py
@@ -1,0 +1,4 @@
+try:
+    import unittest2 as unittest
+except:
+    import unittest